lingosub-server/requirements.txt
fastapi
uvicorn[standard]
celery
redis
pydantic-settings
python-multipart 